pyproject.toml
[project]
name = "march-madness"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.0",
    "numpy>=1.24",
    "pyarrow>=14.0",
    "scikit-learn>=1.3",
    "xgboost>=2.0",
    "optuna>=3.4",
    "rapidfuzz>=3.0",
    "cbbd>=1.20",
    "pyyaml>=6.0",
    "requests>=2.31",
    "joblib>=1.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
